Move test code back out after rebase
